Adds basic logging functionality
Introduces a basic logging system to the engine. This system allows for logging messages with different severity levels (Debug, Info, Warning, Error, Fatal).

The logging system includes a global logger that manages a queue of log messages and writes them to a file and/or the console.
The logger class is a template that can be used to provide context to the logger.